renamed client to service
